Typescriptify WiggleRendering
